requirements.txt
djangorestframework
mixer
Pillow
python-decouple
django-modeltranslation
